Refactor
react-awesome-reveal set to 3.4 according to https://github.com/morellodev/react-awesome-reveal/issues/57
