+++ package.json
@@ -24,7 +24,7 @@
     "next": "^11.0.0",
     "prop-types": "^15.7.2",
     "react": "^17.0.2",
-    "react-awesome-reveal": "^3.8.1",
+    "react-awesome-reveal": "3.4",
     "react-bootstrap": "^1.6.1",
     "react-confetti": "^6.0.1",
     "react-dom": "^17.0.1",
