--- src/html.js
@@ -1,40 +1,0 @@
-/* eslint-disable react/require-default-props */
-/* eslint-disable react/no-danger */
-/* eslint-disable react/forbid-prop-types */
-/* eslint-disable jsx-a11y/html-has-lang */
-import React from 'react';
-import PropTypes from 'prop-types';
-
-export default function HTML({
-  htmlAttributes,
-  headComponents,
-  bodyAttributes,
-  preBodyComponents,
-  postBodyComponents,
-  body,
-}) {
-  return (
-    <html {...htmlAttributes}>
-      <head>
-        <meta charSet="utf-8" />
-        <meta httpEquiv="x-ua-compatible" content="ie=edge" />
-        <meta name="viewport" content="width=device-width, initial-scale=1, shrink-to-fit=no" />
-        {headComponents}
-      </head>
-      <body {...bodyAttributes}>
-        {preBodyComponents}
-        <div key="body" id="___gatsby" dangerouslySetInnerHTML={{ __html: body }} />
-        {postBodyComponents}
-      </body>
-    </html>
-  );
-}
-
-HTML.propTypes = {
-  htmlAttributes: PropTypes.object,
-  headComponents: PropTypes.array,
-  bodyAttributes: PropTypes.object,
-  preBodyComponents: PropTypes.array,
-  body: PropTypes.string,
-  postBodyComponents: PropTypes.array,
-};
