--- src/pages/_app.js
@@ -1,6 +1,0 @@
-import 'bootstrap/dist/css/bootstrap.min.css';
-import '../style/main.scss'
-
-export default function MyApp({ Component, pageProps }) {
-    return <Component {...pageProps} />
-}
